Hide lockscreen visibility switch conditionally
Since this switch is only to hide the power menu while the lockscreen
is secured, it makes sense to only have it visible while the device
is secured. We have many other options both custom and AOSP that work
like this, so is only fitting that we continue this with this switch.